requirements.txt
acdh-django-charts==0.5.2
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-autocomplete-light==3.3.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-tables2==2.0.3
djangorestframework==3.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==1.3.14
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
wincertstore==0.2
